requirements.txt
pandas==2.2.3
ta==0.11.0
alpaca-trade-api==3.2.0
matplotlib==3.10.3
scikit-learn==1.6.1
jupyter==1.1.1
python-dotenv==1.1.0